requirements.txt
aiogram==3.25.0
aiohttp==3.13.3
aiofiles==25.1.0
aiosqlite==0.22.1
asyncpg==0.30.0
fastapi==0.133.1
uvicorn==0.41.0
SQLAlchemy==2.0.47
python-dotenv==1.2.1
pydantic==2.12.5
Jinja2==3.1.6
python-multipart==0.0.20
